move test + fix cloning of eq witness
